refactor: break snapshot.ts <-> snapshot-quality.ts type cycle — Phase 5 prep
snapshot.ts (the foundational snapshot type module; Rect/SnapshotNode, 96
importers, re-exported by contracts.ts) imported SnapshotQualityVerdict from
snapshot-quality.ts, which in turn imports SnapshotNode from snapshot.ts — a
type-level cycle that blocks moving snapshot.ts down into a kernel layer.

Move the SnapshotQualityVerdict type definition into snapshot.ts (it is
referenced by a SnapshotNode field) and re-export it from snapshot-quality.ts so
every existing `from './snapshot-quality.ts'` import is unaffected. The
validation logic (readSnapshotQualityVerdict + const Sets) stays put.

Result: snapshot.ts has zero imports (a clean leaf), unblocking the kernel-seed
move and the eventual snapshot/ domain extraction. No behavior change.

Verified: tsc, oxfmt + oxlint --deny-warnings, fallow audit clean, 369 tests
across utils pass.
